Graph vis package refactor v2
