+++ src/framework/graph-visualiser/rendering/node-diamond/index.ts
