+++ src/framework/graph-visualiser/rendering/node-diamond/shader-frag.ts
