+++ src/framework/graph-visualiser/rendering/node-diamond/shader-vert.ts
