+++ src/framework/graph-visualiser/rendering/node-diamond/utils.ts
