+++ src/framework/graph-visualiser/rendering/node-ellipse/index.ts
