+++ src/framework/graph-visualiser/rendering/node-ellipse/shader-frag.ts
