+++ src/framework/graph-visualiser/rendering/node-ellipse/shader-vert.ts
