+++ src/framework/graph-visualiser/rendering/node-ellipse/utils.ts
