+++ src/framework/graph-visualiser/rendering/node-rounded-rect/index.ts
