+++ src/framework/graph-visualiser/rendering/node-rounded-rect/shader-frag.ts
