+++ src/framework/graph-visualiser/rendering/node-rounded-rect/shader-vert.ts
