+++ src/framework/graph-visualiser/rendering/node-rounded-rect/utils.ts
