+++ src/framework/graph-visualiser/rendering/sigma-settings.ts
