+++ src/framework/graph-visualiser/visual/index.ts
@@ -1,0 +1,9 @@
+export * from "./types";
+export {
+    ILogicalGraphConverter,
+    convertLogicalGraphWith,
+    StudioConverter,
+    shouldCreateNode,
+    shouldCreateEdge,
+    vertexMapKey,
+} from "./converter";
