+++ src/service/query-page-state.service.ts
@@ -9,9 +9,9 @@ import { FormControl } from "@angular/forms";
 import { BehaviorSubject, combineLatest, map, Observable, pairwise, shareReplay, startWith, switchMap } from "rxjs";
 import { DriverAction } from "../concept/action";
 import { GraphVisualiser } from "../framework/graph-visualiser";
-import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/sigma-settings";
-import { newVisualGraph, VisualGraph } from "../framework/graph-visualiser/graph";
-import { Layouts } from "../framework/graph-visualiser/layouts";
+import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/rendering/sigma-settings";
+import { newVisualGraph, VisualGraph } from "../framework/graph-visualiser/visual/types";
+import { Layouts } from "../framework/graph-visualiser/layout";
 import { detectOS } from "../framework/util/os";
 import { INTERNAL_ERROR } from "../framework/util/strings";
 import { DriverState } from "./driver-state.service";
