+++ src/service/schema-state.service.ts
@@ -13,9 +13,9 @@ import Graph from "graphology";
 import { BehaviorSubject, combineLatest, distinctUntilChanged, finalize, first, map } from "rxjs";
 import Sigma, { Camera } from "sigma";
 import { GraphVisualiser } from "../framework/graph-visualiser";
-import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/sigma-settings";
-import { newVisualGraph } from "../framework/graph-visualiser/graph";
-import { Layouts } from "../framework/graph-visualiser/layouts";
+import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/rendering/sigma-settings";
+import { newVisualGraph } from "../framework/graph-visualiser/visual/types";
+import { Layouts } from "../framework/graph-visualiser/layout";
 import { DriverState } from "./driver-state.service";
 import { GraphStyleService } from "./graph-style.service";
 import { SnackbarService } from "./snackbar.service";
